requirements.txt
asgiref==3.6.0
attrs==22.2.0
az-iranian-bank-gateways==1.8.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-rest-swagger==2.2.0
django-schema-graph==2.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
Flask==2.2.3
gensim==4.3.1
hazm==0.9.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.3.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
nltk==3.3
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
platformdirs==3.2.0
pycryptodome==3.17
pydotplus==2.0.2
pyparsing==3.0.9
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
zeep==4.2.1
zipp==3.15.0
